Don't use nested tasks, they don't seem to work
