+++ .gitignore
@@ -9,3 +9,4 @@ __pycache__/
 /*.json
 *.pickle
 kmer_binning.txt.*
+/src/redun
